refactor(query_enhancements): group the PromQL helpers under common/metrics
