+++ src/plugins/query_enhancements/common/metrics/index.ts
@@ -1,0 +1,7 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+export * from './prom_legend';
+export * from './prom_step';
