+++ src/plugins/query_enhancements/common/metrics/prom_legend.ts
