+++ src/plugins/query_enhancements/opensearch_dashboards.json
@@ -5,7 +5,7 @@
   "ui": true,
   "requiredPlugins": ["data", "opensearchDashboardsReact", "opensearchDashboardsUtils", "savedObjects", "uiActions"],
   "optionalPlugins": ["dataSource", "usageCollection"],
-  "extraPublicDirs": ["common/prom_step"],
+  "extraPublicDirs": ["common/metrics"],
   "configPath": ["queryEnhancements"]
 }
